- FIXED: All ships that fire lasers now has turrets drawn on them (beetles lose them when upgraded) - FIXED: Added radar structures to last few levels so the player can't just build rats and win easily - FIXED: Missions/tutorial levels now end after the final explosion, not right when the important object is destroyed - ADDED: A missile firing sound - FIXED: Change the levels of some sounds to match others - ADDED: Refineries now make smoke when refining metal - FIXED: Spider craft now take correct amount of time to rebuild drones - ADDED: The base station spins a little bit now... it looks pretty cool.

+++ trunk/Brogguts/OpenGLEngine/SpiderCraftObject.h
@@ -16,11 +16,11 @@
 	NSMutableArray* droneArray;		// Array of spider drones for this ship
 	int droneCount;					// Current number of drone owned by this ship
 	int droneCountLimit;			// Limit to the number of drones this ship can hold
-    int droneBuildTimer;            // Timer that lets the spider build more drones
+    float droneBuildTimer;            // Timer that lets the spider build more drones
     BOOL droneBayContainment[SPIDER_NUMBER_OF_DRONES];  // containing bools for each drone bay stating whether or not a drone is held in it
     
     int droneBroggutCost;               // Brogguts
-    int droneBuildTime;                 // Seconds
+    float droneBuildTime;                 // Seconds
 }
 
 - (id)initWithLocation:(CGPoint)location isTraveling:(BOOL)traveling;
